React 19 APIs (#133)
* React.promise

* React.useTransition

* React.useActionState

* React.useOptimistic

* React.use

* React.act

* React.useDeferredValue now takes initial value

* ReactDOM: ref cleanup function

* ReactDOM: Resource Preloading APIse

* ReactDOM.useFormStatus

* JS output changes

* Add external for basic support for formAction

* Add `usePromise` for `use(promise)`

`use(context)` seems to exactly replicate the `useContext(context)`
logic, and we want to maximize retro-compability (ie. not switch from
`useContext` to `use` for that under the hood).

simply adding `usePromise(promise)` seems to be the simplest, least
invasive way to add the functionality.

* proposal: make FormData more usable to get values

this **kinda** goes against the zero-cost philosophy, but I don't see a
world where users would not have to reimplement those.

* useOptimistic optionnal updateFn

* dom static prerender and prerenderToNodeStream

---------

Co-authored-by: Matthias Le Brun <bloodyowl@icloud.com>
Co-authored-by: Freddy Harris <freddy03h@gmail.com>